pyproject.toml
[tool.poetry]
name = "newspapers-analysis"
version = "0.1.0"
description = "Project to analyse the change in news oulets coverage in time"
authors = ["drearondov <63360808+drearondov@users.noreply.github.com>"]
license = "UNLICENSED"
readme = "README.md"

[tool.poetry.scripts]
task = "newspaper-analysis.tasks.task_runner:app"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
pandas = "^1.4.3"
numpy = "^1.23.1"
jupyter = "^1.0.0"
spacy = "^3.4.1"
wordcloud = "^1.8.2"
plotly = "^5.9.0"
textblob = "^0.17.1"
kaleido = "0.2.1"
pysentimiento = "^0.5.2"
polars = "^0.15.16"
dash-bootstrap-components = "^1.4.1"

[tool.poetry.group.dev.dependencies]
click = "^8.1.3"
Sphinx = "^5.1.1"
coverage = "^6.4.2"
scipy = "^1.9.0"
myst-parser = "^0.18.0"
boto3 = "^1.24.46"
tqdm = "^4.64.1"
sctriangulate = "^0.12.0"
dash = "^2.8.1"
jupyter-dash = "^0.4.2"
gensim = "^4.3.0"
types-requests = "^2.28.11.13"
boto3-stubs = "^1.26.75"
tabulate = "^0.9.0"
num2words = "^0.5.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_executable = "/Users/drea/.pyenv/versions/3.10.9/envs/nlp-newspapersAnalysis/bin/python"
